\name{csrplus-package}
\alias{csrplus-package}
\alias{csrplus}
\docType{package}
\title{
Methods to Test Hypotheses on the Distribution of Spatial Point Processes
}
\description{
Includes two functions to evaluate the hypothesis of complete spatial randomness (csr) in point processes.  The function 'mwin' calculates quadrat counts to estimate the intensity of a spatial point process through the moving window approach proposed by Bailey and Gatrell (1995).  Event counts are computed within a window of a set size over a fine lattice of points within the region of observation.  The function 'pielou' uses the nearest neighbor test statistic and asymptotic distribution proposed by Pielou (1959) to compare the observed point process to one generated under csr.  The value can be compared to that given by the more widely used test proposed by Clark and Evans (1954).
}
\details{
\tabular{ll}{
Package: \tab csrplus\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-02\cr
License: \tab GPL-2\cr
}
}
\author{
Sarah Smith <smirks2@hotmail.com>
}


